spring-rest-docs-example/src/main/java/com/github/iahrari/springrestdocsexample/web/controller/BeerController.java
package com.github.iahrari.springrestdocsexample.web.controller;

import com.github.iahrari.springrestdocsexample.domain.Beer;
import com.github.iahrari.springrestdocsexample.repository.BeerRepository;
import com.github.iahrari.springrestdocsexample.web.model.BeerDto;

import org.springframework.core.convert.ConversionService;
import org.springframework.http.HttpStatus;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.PutMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;
import org.springframework.http.ResponseEntity;

import java.util.UUID;

@RestController
@RequestMapping("/api/v1/beer")
public record BeerController(
    BeerRepository beerRepository,
    ConversionService conversion
) {
    
    @GetMapping("/{beerId}")
    public ResponseEntity<BeerDto> getBeerById(@PathVariable("beerId") UUID beerId){
        return new ResponseEntity<>(conversion.convert(beerRepository.findById(beerId).get(), BeerDto.class), HttpStatus.OK);
    }

    @PostMapping
    public ResponseEntity<Void> saveNewBeer(@RequestBody @Validated BeerDto beerDto){
        beerRepository.save(conversion.convert(beerDto, Beer.class));
        return new ResponseEntity<Void>(HttpStatus.CREATED);
    }

    @PutMapping("/{beerId}")
    public ResponseEntity<Void> updateBeerById(@PathVariable("beerId") UUID beerId, @RequestBody @Validated BeerDto beerDto){
        beerRepository.findById(beerId).ifPresent(beer -> {
            beer.setBeerName(beerDto.getBeerName());
            beer.setBeerStyle(beerDto.getBeerStyle().name());
            beer.setPrice(beerDto.getPrice());
            beer.setUpc(beerDto.getUpc());

            beerRepository.save(beer);
        });

        return new ResponseEntity<Void>(HttpStatus.NO_CONTENT);
    }
}
